# packages/skyhelper/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/update/ts": {
      "get": {
        "description": "Retrieves information about the current Traveling Spirit",
        "operationId": "UpdateController_getTS",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Traveling Spirit data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "spirit": {
                      "type": "string",
                      "example": "abyss-spirit"
                    },
                    "visitDate": {
                      "type": "string",
                      "example": "15-01-2024"
                    },
                    "index": {
                      "type": "string",
                      "example": "1"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current Traveling Spirit data",
        "tags": ["Game Data Updates"]
      },
      "patch": {
        "description": "Updates the current Traveling Spirit information",
        "operationId": "UpdateController_updateTS",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "spirit": {
                    "type": "string",
                    "example": "abyss-spirit"
                  },
                  "visitDate": {
                    "type": "string",
                    "example": "15-01-2024"
                  },
                  "index": {
                    "type": "string",
                    "example": "1"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Traveling Spirit data updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "spirit": {
                      "type": "string",
                      "example": "abyss-spirit"
                    },
                    "visitDate": {
                      "type": "string",
                      "example": "15-01-2024"
                    },
                    "index": {
                      "type": "string",
                      "example": "1"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Missing or invalid authentication"
          },
          "404": {
            "description": "Spirit not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update Traveling Spirit data",
        "tags": ["Game Data Updates"]
      }
    },
    "/update/events": {
      "get": {
        "description": "Retrieves information about the current active event",
        "operationId": "UpdateController_getEvents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Event data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Days of Bloom"
                    },
                    "startDate": {
                      "type": "string",
                      "example": "15-01-2024"
                    },
                    "endDate": {
                      "type": "string",
                      "example": "29-01-2024"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current event data",
        "tags": ["Game Data Updates"]
      },
      "patch": {
        "description": "Updates the current event information",
        "operationId": "UpdateController_updateEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Days of Bloom"
                  },
                  "startDate": {
                    "type": "string",
                    "example": "15-01-2024"
                  },
                  "endDate": {
                    "type": "string",
                    "example": "29-01-2024"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event data updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Days of Bloom"
                    },
                    "startDate": {
                      "type": "string",
                      "example": "15-01-2024"
                    },
                    "endDate": {
                      "type": "string",
                      "example": "29-01-2024"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update event data",
        "tags": ["Game Data Updates"]
      }
    },
    "/update/quests": {
      "get": {
        "description": "Retrieves current daily quests information",
        "operationId": "UpdateController_getQuests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daily quests data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Daily quests schema object"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get daily quests data",
        "tags": ["Game Data Updates"]
      },
      "patch": {
        "description": "Updates the daily quests information",
        "operationId": "UpdateController_patchQuests",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Daily quests data to update",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quests": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "date": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "images": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "url": {
                                "type": "string"
                              },
                              "by": {
                                "type": "string"
                              },
                              "source": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "last_updated": {
                    "type": "string"
                  },
                  "last_message": {
                    "type": "string"
                  },
                  "rotating_candles": {
                    "type": "object"
                  },
                  "seasonal_candles": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Daily quests data updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Updated daily quests schema object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "401": {
            "description": "Missing or invalid authentication"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update daily quests data",
        "tags": ["Game Data Updates"]
      }
    },
...
  },
  "info": {
    "title": "SkyHelper API",
    "description": "\n      The SkyHelper Bot API provides comprehensive access to Sky: Children of the Light game data and bot management features.\n      \n      ## Features\n      - **Guild Management**: Configure bot settings for Discord servers\n      - **Game Data**: Access shards, events, traveling spirits, and daily quests information\n      - **User Preferences**: Manage user settings and linked roles\n      - **Live Updates**: Configure automated updates for game events\n      - **Reminders**: Set up notifications for game activities\n      \n      ## Authentication\n      Most endpoints require authentication via Bearer token obtained through Discord OAuth2.\n      The token should be included in the Authorization header as: `Bearer <token>`\n      \n      ## Rate Limiting\n      API requests are rate-limited to prevent abuse. Please respect the limits and implement proper retry logic.\n      \n      ## Support\n      For API support, join our Discord server or check the GitHub repository.\n    ",
    "version": "7.7.0",
    "contact": {
      "name": "SkyHelper Team",
      "url": "https://github.com/imnaiyar/SkyHelper",
      "email": "support@skyhelperbot.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/imnaiyar/SkyHelper/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://api.skyhelperbot.com",
...
    },
    {
...
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your Discord OAuth2 access token",
        "in": "header"
      }
    },
...
  }
}